requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
yellowbrick
joblib
streamlit
plotly
Pillow